Remove navbarlistener
